visualización estadísticas

+++ WisteriaDameon/src/CambioDia.java
@@ -25,8 +25,8 @@ public class CambioDia  extends TimerTask {
 		log.debug("Arranques caldera: " + sistema.getArranques());
 		notificaciones.enviar("Arranques caldera: " + sistema.getArranques());
 		
-		// Convierte a horas
-		long tiempo_funcionando = (sistema.getTiempoFuncionando()/3600000);
+		// Convierte a horas en float
+		float tiempo_funcionando = (float)(sistema.getTiempoFuncionando()/3600000);
 		
 		log.debug("Tiempo funcionamiento: " + tiempo_funcionando);
 		notificaciones.enviar("Tiempo funcionamiento: " + tiempo_funcionando);
